bug lp845714
Kind of fixes it, this calls resize on the instance. The space will be
available, but the partition won't be resized to use this space.
Running fdisk -l will show the additional space(after a discussion here
I was told this solution was best for the moment).

Adding myself to Authors

Change-Id: Ice9b12b64b266bd291e7c689114b1172e317e144