Raise InsufficientFreeMemory
Kind of fixes bug 851374 & bug 858679

Raises InsufficientFreeMemory if an instance can't
start because of that. This will cause the normal
instance failure recovery to catch this problem,
set the state, and log the error. This also
removes instance_set_state from db/api.py as that
was causing these exceptions in the first place.

Change-Id: I199aa6900890531b175e28c3b93d8dfb88e135d0