Bug #898290: iSCSI volume backend treats FLAGS.host as a hostname
Change ISCSIDriver to set volume.provider_location during create_export.
This records the location of the LUN, so that nova-compute does not need
to run the iSCSI discovery code itself.

As part of this, include the IP address of the target (--iscsi_ip_address) in
the provider_location.  This means that we don't use volume's host identifier
(which could be an opaque ID) when trying to connect to the iSCSI target --
we use the admin-specified IP address instead.

The string-join to set provider_location is shared with ZadaraBEDriver,
which was doing something similar.  I've brought that into a helper function.

The docstring for the --host flag has been clarified.

Change-Id: I8402da86345e786a46a4d222ad4d8a4449d2bd3f