Add support for Qpid to nova.rpc.
Implements blueprint apache-qpid-rpc.

This patch adds a new implementation of the nova.rpc API that uses Qpid
as the messaging backend.  We first considered trying to add Qpid
support to kombu, but the kombu API did not map very well to Qpid, which
is based on a newer version of AMQP.  It was easier to just map the nova
rpc API to Qpid.

For more information about how to use Qpid with Nova, see this wiki
page:

    http://wiki.openstack.org/QpidSupport

The structure of impl_qpid is largely based on impl_kombu, but adapted
to use the Qpid APIs as necessary.  This patch also factors out some of
the code shared with impl_kombu into nova.rpc.common to try to cut down
on duplicated code.

The unit tests were written from scratch instead of taking advantage
of the common rpc unit tests.  The common unit tests only work if qpidd
is running, but they do pass if enabled.  The unit tests for impl_qpid
instead use mox to mock out the Qpid objects and ensure the right
operations happen on them when the rpc API is exercised.

This patch was a joint effort between myself and William Henry, which is
why he was added to the Authors file in this patch.

Change-Id: Ibacaa956e016ef96f014443074e2a4622e31f090