Add initiator to initialize_connection
Some volumes need to know the name of the initiator that will be
connecting to the iscsi volume.  This adds a call down to the hypervisor
driver to get the ip and the initiator name for the vm before calling
initialize connection. This connection is passed down to the volume
driver so that it can be used to authenticate when the hypervisor
tries to connect to the volume.

 * Adds initiator initialize_connection
 * Makes a call to driver to get initiator name and ip address
 * Gets initiator from openiscsi for libvirt
 * Gets initiator from config for xenapi
 * Add tests for the driver calls
 * Fixes bug 924461

Change-Id: I5b6a2dd84560c7f7b447571e0abf0993e5512ca0