Remove Hyper-V support
HyperV has been unmaintained for several releases now. The unit tests
are superficial, we have no way to test it, noone has stepped forward
to maintain it, and for a very long time, we've not had any reports that
it works. Furthermore, many improvements have been made across other
hypervisor drivers that have not been done in the HyperV driver, so even
if it worked, it would only expose a subset of the functionality that
the other drivers do.

Implements blueprint hypervsectomy.

Change-Id: I3160915816a305d7d6bc34790f9deac459b562a9