Consistently update instance in nova/compute/manager.py
The code used a mix of self.db.instance_update and self._update_instance,
so settle on self._update_instance. Also, the code used a mix of
self._update_instance and self._set_instance_error_state, so settle on
self._set_instance_error_state. This also changes the code to not clear
task_state in all cases, leaving it set for debugging purposes.

Change-Id: I5c841ebfb60ba4acf62382060e416ea780bd66ba