nova-manage: allow use of /32 IP range
The current code does not allow to add an IP range which is /32, because
netaddr considers that there's no host in such a network. While this is
probably debatable, it disallow to add specific IP address in a pool, which
is really handy and was possible before.
This patch fix that by treating the /32 subnet as a special case.

Change-Id: I38685e6f1a3541519d1f2a9ec8d3b00dd522b44a