Include launch_index when creating instances.
Fixes bug 934534.

The launch_index column in the database is used as the ami-launch-index value
provided by the metadata service.

If launch_index is not specified it defaults to 1. This is incorrect for the
single instance case (should be 0) and doesn't correctly handle multi-instance
starts.

This branch changes the default launch_index to 0 (to handle the single
instance case) and adds the launch_index in the schedulers for multi-instance
starts.

Change-Id: Ifc45abf4cd9f50f732ba4a4b68c0a6242a6c9710