bigger-than-unit test for cleanup_running_deleted_instances
My team is going to lambaste me for this, but the
cleanup_running_deleted_instances was missing a
unit test. This lead to a bug in the refactor.
This patch fixes both issues; however
utils.temporary_mutation makes it hard to write a
proper unit test, this one hits the DB.

Change-Id: I93a595ffce4f17261f18a52d4c2a11434653a630