get_instance_uuids_by_ip_filter to QM
Adding a get_instance_uuids_by_ip_filter to the
QuantumManager. Also adds the supporting calls in
melange_ipam_lib an connector to support this. It
currently only works with Melange, I still need to
add a conditional that falls back on the original
implementation in the FlatManager if not using
Melange.

update: rebase

Change-Id: I9468de130c6cd1e0520f9d92c43f98aa6b36dce6