Auto-confirming resizes would bail on exceptions
Fixes bug 971939

Fixes a few things.  If instance went to error, we should set the
migration to error so we don't retry.  Also, if we get any other
exceptions when trying to confirm resizes, we should not bail out of the
loop.  Also: be nice to other greenthreads and yield on each iteration.

Includes complete unit test for xenapi poll_unconfirmed_resizes which
didn't exist prior to this patch.

Change-Id: I6f7c1ccfd4be0f1497727d817130a821902d134d