Remove ugly instance._rescue hack
The XenAPI driver would use _rescue to note when an instance was in
the process of being rescued. The existance of this attribute would
modify the 'name' property to append '-rescue' to the generated name.

Among the problems with it, it wasn't saved anywhere (with resulting
code to set it to False being unnecessary) and it led to "magic" code
(see instance.name) that was confusing without understanding the hack.

Change-Id: I1a91e4efdf69d4c1fba07658419f7bb30b733b5a