Refactor libvirt config classes for representing CPU models/features
The previously added (but not used) LibvirtConfigCPUTest class
was too inflexible. It did not distinguish between parts of the
XML schema which applied to both capabilities & domain XML,
vs those which only applied to the domain XML. By representing
features as a plain string it did not allow for setting attributes
on feature flags like policy.  This change replaces the single
LibvirtConfigCPUTest class with 4 new classes:

 - LibvirtConfigCPUFeature - base class for defining CPU features
 - LibvirtConfigCPU - base class for defining CPU models
 - LibvirtConfigGuestCPUFeature - extension for setting the guest
   specific feature policy
 - LibvirtConfigGuestCPU - extension for setting the guest specific
   match policy, and allowing use of host CPU model passthrough

Fixes: bug #1003373
Implements: blueprint libvirt-xml-cpu-model
Change-Id: I0aa0ddfb86cf8b89b2e4dcc95e21bdca304bd6b3
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>