Adds default fall-through to the multi scheduler. Fixes bug 1009681
This offers a default scheduler driver that will be used by the
multischeduler for requests that are not destined for the
compute or volume topic. This is useful for openstack extensions
that want to use the nova-scheduler to schedule requests to
compute nodes but provide their own manager and topic.

Change-Id: I753e87fc8979fd0b62529974f00cf11fa55b3d63