Ensure system metadata is sent on new image creation.
System metadata is useful for billing and notification purposes.
Sadly, this data was getting lost on new instances created from
snapshots of old instances.

This patch ensures the system metadata of the old instance is
stored in the snapshot for use with creating new instances.

Additionally, it strips the "image_" prefix from any system
metadata keys (as required).

Change-Id: I9fff55c0e9f106afcded1865c7a47bc96ba507a7