Add extensions for flavor swap and rxtx_factor
The swap and flavor attributes of a flavor are not in the spec. This
moves the properties so they are generated by extensions. The output
will not be changed if all extensions are enabled (by default), but
we now have a way to document these extra attributes and disable them.

DocImpact

Change-Id: Iee1cb1b1ee4116a38b90db581c38468d3d92afaf