Fix race condition in CacheConcurrencyTestCase
Fixes bug 1052641

Recent changes made to fix bug 1035426 ended up creating a race
condition in the tests for CacheConcurrencyTestCase. Since two
greenthreads are used, the changes to utils.synchronized() can
cause a race where one thread is deleting the lock directory out
from under the other thread.

An unrelated bug caused this race condition to be ignored since
the greenthreads were never waited on and as a result, the exceptions
never had a chance to be raised in the test case.

Change-Id: I86b45414bbacd12bd3f5af323d26e0861c354ba8