Fix solidfire unit tests
A couple of tests attempted to reimplement assertRaises(), but did so
badly. self.fail() will raise AssertionError, which is a subclass of
Exception, so even if the tested exception wasn't raised, it would
silently treat it as a success.

This buggy testing code hid a failure to properly test the intended
functionality in test_delete_volume_fails_no_volume.

Change-Id: I4e31bf6cb3b042aa2d9ba2a5ccaf40319d0fc124