remove deprecated connection_type flag
Fixes bug #1025712

connection_type was deprecated in Folsom and should now
be removed in Grizzly as early as possible to shake out
any fallout. This removes all references to it, changes
the config sample and tests appropriately.

Remove old default fixtures for flavors that specify
ephemeral disks if the connection_type flag was set to
'libvirt'

Change-Id: I8af831600a1931ae92c6d06c5105bd1bd81debe3